fix referrers, catalog
This fixes the referrers and catalog APIs in both ociclient and ocimem.

Also expose ErrorIter in ociregistry as it seems generally useful.

Add some tests for referrers and catalog inside the conformance testing
module, using the oras-go client to give us some assurance that
we're adhering to generally understood conventions.