ociunify: fix flakiness
The body returned from a `net/http` request is still controlled by
the lifetime of the context that was passed in to the request.

This meant that starting two requests and cancelling them
before returning the result (as `runRead` currently does) won't
work because returned reader gets a `context canceled` error when
returning.

Fix it by tying the lifetime of the context to the returned reader.

Also make it explicit in the docs that the returned reader lifetime
is tied to the context passed to the request. This fits nicely with
the current behaviour of `BlobWriter` which follows the same pattern.