ociregistry/ociclient: do not use single-post upload
As outlined in [this Go issue](https://github.com/golang/go/issues/63152)
and [this docker registry issue](https://github.com/distribution/distribution/issues/4065),
failing over from single-post to post-then-put is problematic because
we can't guarantee that the body won't be consumed.

We could implement some logic to cache the request body
until the 100-continue response arrives (I have already made a PoC
to verify this), but a simpler solution, if somewhat less efficient, is always
to use the post-then-put method, which doesn't require that failover.