ociregistry/internal/conformance: re-add replace directive
It turns out that keeping it in serves a purpose.
Without the replace and with the "all zeroes" version,
`go mod tidy` does not work since it encounters an invalid version.

We could replace the invalid version with a published pseudo-version,
but that doesn't really make sense either, since we always want to test
the local version in the same repository checkout.

All in all, this module is special - it's only for testing.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Id9935f241187aa78edac2f7018901f86bbc324e4
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1170071
Reviewed-by: Paul Jolly <paul@myitcv.io>
Reviewed-by: Chief Porcuepine <chief.porcuepine@gmail.com>