ociregistry/ociclient: don't panic on nil Options in New
The check on Insecure was being done on opts rather than opts1,
so passing a nil *Options caused a nil pointer dereference panic.

We could fix that expression to use opts1, but the footgun would remain.
To avoid having the two names in scope and prevent further human error,
use opts consistently instead.

We should be able to trust the compiler to do well enough with the
allocation as well as the nil checks here.
And even if it doesn't, performance is not crucial in this function.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I0b0070ca9ad6a0c04d71e15b7047304e9d4dc209
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1170347
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Roger Peppe <rogpeppe@gmail.com>