ociregistry: define and use ErrRangeNotSatisfiable
Since ociserver is meant to be stateless, it can't possibly know
whether a chunked upload request should return a status code 416
when the range in the request is invalid.

Currently, the code relies on the next OCI registry interface
to report what was the last offset via the Size method;
using ociclient, this results in an extra GET request.

Instead, we want to start relying on the final OCI implementation,
like ocimem, to report these range errors.
After all, it's only the registry holding the content
which needs to keep the state about what bytes have been uploaded.

To do this, we need to thread "range not satisfiable" errors through
ociserver and ociclient, meaning they need to survive HTTP round-trips.
Both ociserver and ociclient already know how to do this for spec errors
since they are uniquely identified via string codes.
The spec doesn't define this range error nor give it a string code,
so we define it as a regular error and identify it by http status code.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: If487232f7375389b6fcbd03fa3a59f951d2171f3