ocimem: start using ErrRangeNotSatisfiable
That is, if a chunked upload is resumed at an offset that isn't -1,
the first write will result in an error if the offset doesn't match
with where the last chunk left off.

Note that this check happens at the first Write,
as opposed to in the PushBlobChunkedResume call,
because ociserver implements closing a chunked reader as follows:

    w, _ := client.PushBlobChunkedResume(ctx, repo, id, 0, 0)
    w.Commit() // sends a PUT with an empty body, since we didn't write
    w.Close()

In this particular case, which is allowed by ociregistry.Interface,
we don't want the offset to be checked at all.
Per the spec, the closing PUT does not include a Content-Range header,
so ociserver can't know what the last offset was without keeping state.
We want the code above to result in a PUT without any error or GET,
so the easiest way is to only check the offset at the first write.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I464c22b165cdcde80ba2cffa9091fd095c53e9b8