ociregistry/ociserver: forward offsets to PushBlobChunkedResume
This means that the majority of PATCH and PUT requests will no longer
result in an extra leading GET request to get the previous offset.

Initially I had written this to be stricter following the spec;
PATCH would always require a Content-Range,
and the three kinds of PUT (POST-PUT, chunked PUT without body,
chunked PUT with body) would be separately handled depending on
headers like Content-Type and Content-Range.

However, this resulted in broken conformance tests.
For example, the fact that the chunked PUT without body
does not require a Content-Type request header
does not mean that using such a header is forbidden by the spec.
In fact, one of the conformance tests does this, surprisingly enough.

We still explain what the three modes in the PUT handler are,
but they are all handled the same way in terms of the logic.
We ensure that PushBlobChunkedResume never performs a GET
by ensuring that the given offset is never -1.

Per input from Roger, ociserver and ociclient now share the code
to parse and print range strings, to deduplicate code.
The server needs a little bit of logic on top of that
to return errors and extract the chunk size as well.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I01401fbdad867050df04bbcda07509d703f477c6
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1170715
Reviewed-by: Roger Peppe <rogpeppe@gmail.com>
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>