ociregistry/ociclient: properly set Content-Range in PATCH requests
It was using the chunkInProgress field to compute the end of the range,
which hasn't been used for some time now.
Effectively, the range always said we were uploading exactly one byte,
even though the Content-Length header was being set correctly.

We didn't notice this in production because it seems like registries
didn't really care about this discrepancy.
Our ociserver package didn't care either, but now it does complain
if both Content-Range and Content-Length are set and disagree.

This fixes the proxy tests with chunked uploads.
Since ociserver's handling of a PATCH derives what chunk size to use
from the request's Content-Range, our client giving it a range of 0

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I2b79d156557204d0255bd38ea2e02384d9226467