ociauth: rename from ociauthconfig
We're going to put all the auth-related code into the `ociauth` package,
so start by moving the authfile-parsing code there.

Change some of the naming, as "Auth" prefixes seem redundant if
they're in the context of an `ociauth` package.

Also make `Config` into an interface so that it can become a pluggable
value in the future if we choose to use different kinds of config source.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ie6219acf29cf5ee19dcd507d776ced24260f266d