ociauth: add Scope
This provides a primitive that can be used for constructing authz
scopes. The data structure we use is reasonably efficient in storage
overhead and for union operations. Arguably it's overkill for
the time being when in practice we'll probably only be using
scopes with a small number of elements, but perhaps it's
worth keeping so that we can easily switch to using
larger scope sets for registries that support them.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Idea30b850face768c61e5db48369877fb5befe9a