ociauth: preserve scope strings
It seems that some registries are picky and require exactly
the same scope to be asked for that was returned in
the Www-Authenticate challenge. We can allow for
that while still preserving reasonable scope arithmetic
by preserving the original string that a scope was created from.

As long as the contents of the scope remains the same, its
string form will remain unchanged.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ic98910ce4abd66641d2601d60d3e1dab87225440