ocifilter: make Sub pass auth scopes along correctly
The scopes passed in by the caller will refer to repositories
from the caller's perspective, but the underlying registry
implementation expects them from its own perspective,
so transform them as they're passed down.

Also rename the source files, as "prefix.go" was a hangover
from the previous API when the call was named `Prefix`.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I6d5f18faf7e18abb47dc2a0d4cda140375cda13e