ociregistry/internal/exp: vendor exp/{slices,maps} until go1.22
Vendored from golang.org/x/exp@v0.0.0-20231006140011-7918f672742d, with
minimal changes to create a working version:

* Removing test files.
* Removing go:generate directives.
* Updating import paths.

Given ociregistry is a core dependency of cuelang.org/go, we should
avoid dependencies on experimental, unstable packages/modules, because
this can cause problems for consumers of ociregistry and things that
depend on it, cuelang.org/go included.

This CL prepares for dropping a dependency on golang.org/x/exp in a
later CL.

When we only support go1.21 and go1.22 we can drop this new package and
instead depend on slices and maps in the standard library.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: Iddf1b2aa14e22a3dd7f7530f72c89eaac4423e8f