ociregistry/ocitest: fix for recursive references
The algorithm didn't work with multiple levels of reference
because adding a new manifest to the "required" list didn't
count as progress. This fixes it so it works.

Also add a couple of top-level functions that don't require
a `*testing.T` to make it possible to use `ocitest` from
arbitrary non-test code.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ia996f8deaa1514ce87adcc0707a369d30aca3013
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1176883
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>