ocimem: add some tests and fix DeleteTag
We want some specific checks for ocimem behaviour over and above the
conformance tests. We also fix the behavior of `DeleteTag` so that
it just deletes the tag, not the manifest that it points to, which
seems to be what's expected (c.f. https://stackoverflow.com/questions/71576754/delete-tags-from-a-private-docker-registry).

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Iee9e1466445e2a0f54793b23304839c47e698d98
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1176884
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>