ociauth: implement `GetBody` TODO
This is suggested by the godoc for `Authorizer.DoRequest`, but was not implemented in the `ociauth`-provided implementation.

This is especially useful for endpoints like `PushManifest` in `ociclient` which uses `bytes.NewReader` and is thus restartable (and Go's `http.NewRequestWithContext` sets `GetBody` automatically for us, in that case).

Without this, I was getting `net/http: HTTP/1.x transport connection broken: http: ContentLength=6415 with Body length 0` when `ociauth` would retry my `PUT` request after collecting the necessary authentication token, and I've verified that this does fix it.

Signed-off-by: Tianon Gravi <admwiggin@gmail.com>