doc: add CONTRIBUTING guide
Based on the CUE contributing guide. Most importantly at this stage it
clarifies the requirement around DCO.

We already enforce requiring of DCO as part of CI in our trybot
workflow. After landing this change we will add the DCO GitHub App to
the GitHub repository to help make obvious for PR contributors that DCO
is required.

Signed-off-by: Paul Jolly <paul@myitcv.io>
Change-Id: I8c9419f8b8c0392b5c72bcc7f8efb118bddb109a