ociregistry/ociclient: limit page size
We provide the ability to limit the page size requested by ociclient,
and change the default to 1000 to conform to the limits imposed
by the AWS ECR registry.

Also add the ability to fail in the same way to ociserver so
we can test this behavior, and in passing remove some unused
code from ociclient.

Updates #6

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ibaa32575049f1583d2ea42d9a0ca9530b2917490
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1177234
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>