ociclient: cope with omitted digest in tag GET response
The AWS ECR registry does not return a Docker-Content-Digest
header in its response when doing a GET of a tag.
Technically this is conformant because the distribution
spec says only that this "should" be returned in this case.

However it is problematic because we need the digest to return
in our API (because `BlobReader` exposes the digest).

So we work around this by either reading the manifest if it's
not too large, or by issuing a HEAD request if it is (technically
the digest is also optional in the HEAD response, but in practice
the ECR registry does return the digest in that response;
other clients treat it as mandatory and fail if not present,
and there's no other way to resolve a tag to a digest).

Also add a way to get ociserver to emulate this quirk so we
can test it.

Updates #6

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I1694498a589fd029fc2a77e78ac864fd1ba9bba8