ociregistry: breaking change: use stdlib-compatible iterator
When `ociregistry.Interface` was designed, the Go iterator proposal was
in very early stages, but now it's implemented in Go 1.22 and will
become fully available in Go 1.23. By using the standard signature, we
can make our iterators compatible with for loops in the future without
actually needing to rely on any new Go features.

This seems like a good thing, and as a bonus the resulting code is
somewhat more concise too.

Unfortunately this breaks the API, but I think it's worth doing anyway
at this early stage.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I70e7c54e44b2d02fd1d6ae58ec94a32ef8a3efd0