ociregistry: breaking change: add paging support for list endpoints
This API-breaking change adds the ability to start at an arbitrary point
in the list results, as described in the specification, by adding a
`startAfter` argument to the `Tags` and `Repositories` methods.

It also changes the client to properly understand Link headers in
responses, and the server to generate those links, as that is the
preferred method of iterating when available (and the only method when
iterating through Referrers results).

Note that the `_catalog` endpoint is not actually documented in the
specification, but existing client implementations demonstrate that the
API is very similar to that for tag listing.

Fixes #3.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Id97a6f58ac073b462f2540cb0e34142897f9b4ba
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1177300
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>