ociregistry: add iterator test
This is ensures that the new iterator type is in fact
compatible with the range-over-func experiment.

It's not tested in CI as yet, but I've made sure it
works locally, and the likelihood of breakage seems
very small given that the iterator type is encapsulated
in `ociregistry.Seq`.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I17ea71d492cee91c5f5b6c617110104d711bec13
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1177541
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>