ociregistry/ociauth: allow passing in explicit environment
This makes it possible to control exactly what variables are read from
the surrounding environment. We use the same conventions used by
`os/exec` for passing in the environment, as we sometimes actually
have to invoke an helper executable.

I decided against making it a function with a config struct because
hopefully there won't be any need to add more parameters in the futue.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I602ec211c323e2ae1db9937a9dc85533de1c43b3
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1177600
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>