ociregistry/ociclient: use only a single HTTP client field
This is a backwardly incompatible change, but hopefully worthwhile in the
long run.

Currently `ociclient.Options` contains two fields which are used to
make HTTP requests: `HTTPClient` and `Authorizer`. This is awkward
because there is no way to use both, and the precedence between the
two is not obvious. Given that it's easy to make an adaptor from a
regular HTTP client to the interface required by `ociauth`, this CL
simplifies the API to provide only a single way to provide the HTTP client.

As pointed out by @mvdan, the `Authorizer` name doesn't really fit its
usage, which is really just an HTTP client that's passed some additional
auth information, so clean all this up in one fell swoop by renaming
`ociauth.Authorizer` to `ociauth.ScopedHTTPClient`, and renaming
`ociclient.Options.Authorizer` to `ScopedHTTPClient`. With that change in place,
the name `ociauth.HTTPDoer` seems inconsistent, so rename it to
`HTTPClient`.

To summarise the changes made:

- rename `ociauth.Authorizer` to `ociauth.ScopedHTTPClient`
- rename `ociauth.HTTPDoer` to `ociauth.HTTPClient`
- remove `ociclient.Options.Authorizer`
- change the type of `ociclient.Options.HTTPClient` to `ociauth.ScopedHTTPClient`.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I89904e1872d1df39623a274c64d916776a52a704