ociclient,ociauth: breaking change: use standard http.Transport
This is a backwardly incompatible change, but hopefully worthwhile in
the long run.

Currently `ociclient.Options` contains two fields which are used to make
HTTP requests: `HTTPClient` and `Authorizer`. This is awkward because
there is no way to use both, and the precedence between the two is not
obvious.

Originally I argued against making the underlying `Authorize` method
the same as a standard HTTP transport because the contract is clearer
if the required scope is a mandatory argument. However, I've come
around to the point of view that using the standard interface is
a better approach despite that.

We now have two context values: one for the required scope (I've actually
associated that with a `RequestInfo` struct so there's space later to
add more fields such as the actual request parameters), and one for
the "desired scope" (that one was already present).

To summarise the changes made:

- use `http.RoundTripper` instead of `ociauth.Authorizer`
- delete `ociauth.Authorizer`
- add `ociauth.ContextWithRequestInfo`
- add `ociauth.RequestInfoFromContext`
- remove `ociclient.Options.HTTPClient`
- remove `ociclient.Options.Authorizer`
- add `ociclient.Options.Transport`

Also, as a minor drive-by, prevent `ociclient.New` from side-effecting
the contents of the `Options` parameter it's passed.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I89904e1872d1df39623a274c64d916776a52a704