ociregistry/ociauth: do not attempt to close nil request bodies
The net/http.RoundTripper docs do say that RoundTrip must always close
the request body, but the net/http.Request.Body docs say:

    For client requests, a nil body means the request has no
    body, such as a GET request. The HTTP Client's Transport
    is responsible for calling the Close method.

That is, client requests are allowed to have a nil body,
and all other RoundTrip methods use a nil check before calling Close.

Marcel ran into a panic when testing the cuelang.org/go/cue/load package
and I think the reason we hadn't caught this in our other tests
is because our RoundTrip only calls Close if it fails early,
before it can hand off that responsibility to another transport.

Add a test which sets up our RoundTrip to fail early,
send it a request with a nil body, check that we used to panic,
and fix the panic with a nil check.
Note that this test is different, as all other tests expect to succeed.

While here, I noticed that another needBodyClose assignment was dead,
as we already set the boolean to false unconditionally further up.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I794ec71f7e220217e01ce6b67c40087cd33a4ef8