ociregistry/ociauth: do not fail when non-host-specific helper is not found
This changes the `ociauth` logic so that it can cope with the situation
that a helper executable is not found even though the helper
is mentioned in the configuration file.

This works around the fact that some tools
[add a helper automatically](https://github.com/docker/roadmap/issues/509)
even though that helper might not be installed.

The docker command goes a bit further and ignores all errors
from a helper command, but we think that's going a bit too far.
If someone has configured a helper for a particular registry,
or if a helper produces invalid JSON, it seems to us like
that should be a hard error.

For https://github.com/cue-lang/cue/issues/2934.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Id0cab8d6915d9695c4a52a49a72c4a348086f8e1