ociregistry: drop internal/exp
Now that cuelang.org/go requires Go 1.21 or later,
we can also drop support for Go 1.20 in the oci repository.

The slices package can be replaced directly by std now.
The maps package is not as simple, as the Keys API was delayed
until iterators were stabilized, so for now, copy one function,
which is significantly easier and smaller too.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: Id7fc6d63c7b44f5ee5e86184d60b247c7c24852c