ociregistry: do not use goexperiment build tag
The build failed for me on Go tip as of mid March:

    $ go test
    # cuelabs.dev/go/oci/ociregistry [cuelabs.dev/go/oci/ociregistry.test]
    ./iter_test.go:15:22: cannot range over SliceIter(slice) (value of type Seq[int]): requires go1.23 or later (-lang was set to go1.21; check go.mod)
    ./iter_test.go:25:25: cannot range over ErrorIter[string](err) (value of type Seq[string]): requires go1.23 or later (-lang was set to go1.21; check go.mod)
    FAIL	cuelabs.dev/go/oci/ociregistry [build failed]

I reckon the "or" build tag expression should work in principle,
but currently it doesn't due to a design quirk.
I've raised https://go.dev/issue/66399 to suggest an upstream change.
For now, un-break the build; testing this file via Go tip is enough.

Signed-off-by: Daniel Martí <mvdan@mvdan.cc>
Change-Id: I017a6437ee6d877ffa11de8be808f58da63fd7fc