ociregistry: improve errors with respect to HTTP status
Currently the standard wire representation of an error is
duplicated across both client and server, and there is
no way for:
- a client to access the actual HTTP status of a response
- an `Interface` implementation to cause the `ociserver` to return
a specific HTTP error code for a error code that it doesn't
know about.

This change addresses that by moving the wire representation
into the top level `ociregistry` package, splitting `HTTPError`
into its own type, and making both `ociclient` and `ociserver`
aware of them.

Fixes #26.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I5a79c1c6fec9f22c1f565830d73486b406dd181d