ociregistry: implement MarshalError and use it in ociserver
This provides a standard way for code to convert
from a regular Go error to the form appropriate for OCI
error response bodies.

It seems to make sense to put this inside the top level
ociregistry package because the JSON-oriented
`WireError` types are already implemented there, and
the logic for stripping error code prefixes is directly related
to the logic in the same package that adds them.

Also in passing fix an unintended use of the deprecated
`io/ioutil` package to use `io` instead.

Also add a test for error stuttering to `ociclient` and fix
the code to avoid stuttering the HTTP status code
as exposed by that test.

Fixes #31.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I22408dd3320b73bc52f37181987f79003dbaa115