ociclient: require digests for descriptors
In PR #29, I suggested that we should verify that the
tests fail when the fix was not applied, but the tests
did not fail in that case.

It turns out that's because the client was being a bit
too slack about requiring digests in responses, meaning
that the client succeeded even when it should have failed.

This change makes `ociclient.descriptorFromResponse`
a little more stringent about digests, and I've explicitly
verified that when the changes in #29 are reverted,
the tests do in fact fail as expected.

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: Ic7352b572593a60dc2209193fdd64bf2e4235af9