ocimem: better Referrers implementation
We support filtering by artifact type as mentioned in the TODO,
but also we provide information on artifact type and annotations
to the client, as specified in the spec.

Also update to a new version of the conformance tests,
which actually adds some testing of the Referrers API,
thus potentially making our existing tests of that API redundant
but we'll leave them there anyway as they're doing no harm for
now.

Fixes #39

Signed-off-by: Roger Peppe <rogpeppe@gmail.com>
Change-Id: I8a1ed01608cfbb40ebe3d890deea2fdc031c8027
Reviewed-on: https://review.gerrithub.io/c/cue-labs/oci/+/1218299
TryBot-Result: CUE porcuepine <cue.porcuepine@gmail.com>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>